requirements.txt
# Agent Skills Integration Requirements

# Core dependencies
agentscope>=0.0.1
PyYAML>=6.0
python-dotenv>=1.0.0

# Optional dependencies for skills
# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Data Analysis
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
scrapy>=2.7.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: For enhanced features
# openpyxl>=3.1.0  # Excel support
# lxml>=4.9.0      # XML/HTML parsing
# pillow>=9.0.0     # Image processing
# jupyter>=1.0.0    # Notebook support
